adb: Add ability to specify device path For manufacturing and testing, there is a need to talk to whatever device is connected to a given port on the host.  This change modifies adb's "-s" option to take either a serial number or a device path.  The device paths of the connected devices can be listed using "adb devices -l" whose output will resemble:
    List of devices attached
    016B75D60A00600D	usb:2-5	device
    3031D0B2E71D00EC	usb:1-4.3	device

The second column lists the device paths.  If the -l option is
not given, the output from "adb devices" will be the same as
it used to be (i.e. the paths will not be printed).

The device path can also be obtained with the get-devpath
command:

    $adb -s 3031D0B2E71D00EC get-devpath
    usb:1-4.3

Note that the format of the device paths are platform dependent.
The example above is from Linux.  On OS-X, the paths will be
"usb:" followed by hex digits.  For other platforms, the device
paths will be printed as "????????????" and the -s option will
not be able to select a device until someone implements the
underlying functionality.

Change-Id: I057d5d9f8c5bb72eddf5b8088aae110763f809d7
Signed-off-by: Scott Anderson <saa@android.com>

Manually picked from https://github.com/CyanogenMod/android_system_core/commit/e109d266c12c5f537d429ca4b892f2719e02c2da